requirements.txt
Django>=5.0
python-dotenv
requests
django-allauth
PyJWT
cryptography
